Added donation controllerAPI

+++ src/main/java/com/jkt/donatebox/repository/NgoRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface NgoRepository extends JpaRepository<NGO,Integer> {
+public interface NgoRepository extends JpaRepository<NGO,Long> {
     @Query("Select n from NGO n where n.address.city = ?1 AND n.type = ?2")
     List<NGO> getNGOByCity(final String city, final NGOType ngoType);
 }
